requirements.txt
asgiref==3.5.0
Django==4.0.4
djangorestframework==3.13.1
et-xmlfile==1.1.0
openpyxl==3.0.9
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
